+++ AzeriteUI/WoW11/ActionBars/PetBar.lua
@@ -29,5 +29,7 @@ if (not ns.WoW11) then return end
 
 -- Do not automatically load this,
 -- let the primary actionbar module do it.
-local PetBarMod = ns:GetModule("PetBar")
+local PetBarMod = ns:GetModule("PetBar", true)
+if (not PetBarMod) then return end
+
 PetBarMod:SetEnabledState(false)