+++ AzeriteUI/WoW11/ActionBars/StanceBar.lua
@@ -29,5 +29,7 @@ if (not ns.WoW11) then return end
 
 -- Do not automatically load this,
 -- let the primary actionbar module do it.
-local StanceBarMod = ns:GetModule("StanceBar")
+local StanceBarMod = ns:GetModule("StanceBar", true)
+if (not StanceBarMod) then return end
+
 StanceBarMod:SetEnabledState(false)