+++ AzeriteUI/WoW11/Misc/Tooltips.lua
@@ -29,7 +29,9 @@ if (not ns.WoW11) then return end
 
 -- Do not automatically load this,
 -- let the primary actionbar module do it.
-local Tooltips = ns:GetModule("Tooltips")
+local Tooltips = ns:GetModule("Tooltips", true)
+if (not Tooltips) then return end
+
 Tooltips:SetEnabledState(false)
 
 Tooltips.OnInitialize = function(self)